requirements.txt
setuptools
wheel
aiocoap
tornado
pyserial-asyncio
pyroute2
netifaces
pycrypto
